Altered removed unwanted file
